Used notebook utils in notebooks
