+++ docs/examples/conditioning-on-zero-residual.md
@@ -37,7 +37,6 @@ plt.rcParams.update(notebook.plot_config())
 if not backend.has_been_selected:
     backend.select("jax")  # ivp examples in jax
 
-
 config.update("jax_enable_x64", True)
 config.update("jax_platform_name", "cpu")
 ```
