+++ probdiffeq/doc_util/__init__.py
@@ -1,1 +1,7 @@
-"""Documentation-related utility functions (work-precision diagrams, etc.)."""
+"""Documentation-related utility functions (work-precision diagrams, etc.).
+
+The modules in here are not part of the public API.
+
+They are only used for the documentation (e.g. examples, benchmarks).
+The dependencies for this module are the "example" dependencies.
+"""
